Extract test fixtures to conftest module
